create rol-user-row
