+++ resources/views/livewire/internal/users-index.blade.php
@@ -45,77 +45,7 @@
                       <tbody class="bg-white divide-y divide-gray-200">
                         <!-- ROWS  -->
                         @foreach($users as $user)
-                          <tr  style="height:auto;">
-                            <td class='px-6 py-4 text-sm font-medium text-right align-middle hard_left iconitos whitespace-nowrap'>
-                              @if(!$editable)
-                                <div>
-                                  <div class="px-2 btnDiv">
-                                    <button  wire:click='edit' class="w-full mt-1 font-bold text-pink-700 bg-pink-300 rounded-md deleteBtn button button-small edit focus:outline-none">
-                                      Editar
-                                    </button>
-                                    <button   class="w-full mt-1 font-bold text-pink-700 bg-pink-300 rounded-md deleteBtn button button-small edit focus:outline-none">
-                                      Eliminar
-                                    </button>
-                                  </div>
-                              @else
-                                  <div >  
-                                    <button class="w-full mt-1 font-bold text-pink-700 bg-pink-300 rounded-md deleteBtn button button-small edit focus:outline-none">
-                                      Guardar
-                                    </button>
-                                    <button wire:click='cancel'  class="w-full mt-1 font-bold text-pink-700 bg-pink-300 rounded-md deleteBtn button button-small edit focus:outline-none">
-                                      Cancelar
-                                    </button>
-                                  </div>
-                                </div>
-                              @endif
-                            </td>
-                            <td class="px-6 py-4 next_left place-content-center whitespace-nowrap" style="position:relative;" data-field="id">
-                              {{$user->id}}   
-                            </td>
-                            <td class="px-6 py-4 whitespace-nowrap" style="position:relative;" data-field="name" >  
-                              {{$user->name}} 
-                            </td>
-                            <td class="px-6 py-4 whitespace-nowrap" style="position:relative;" data-field="email" >
-                              {{$user->email}}
-                            </td>
-                            <td class="px-6 py-4 whitespace-nowrap" style="position:relative;" data-field="status" >
-                            @if(!$editable)
-                                {{$user->status}} 
-                              @else
-                                <select>
-                                  <option value="Admin" {{ $user->status == 'activo' ? 'selected' : '' }}>
-                                    Activo
-                                  </option>
-                                  <option value="User" {{ $user->status == 'inactivo' ? 'selected' : '' }}>
-                                    Inactivo
-                                  </option>
-                                </select>
-                              @endif
-                            </td>
-                            <td class="px-6 py-4 whitespace-nowrap" style="position:relative;" data-field="role" >
-                              @if(!$editable)
-                                {{$user->role_name ? $user->role_name : 'SIN ROL ASIGNADO'}} 
-                              @else
-                                <select>
-                                  <option value="">
-                                    Asigná el rol
-                                  </option>
-                                  <option value="Admin" {{ $user->role_name == 'Admin' ? 'selected' : '' }}>
-                                    Administrador
-                                  </option>
-                                  <option value="User" {{ $user->role_name == 'User' ? 'selected' : '' }}>
-                                    Usuario
-                                  </option>
-                                </select>
-                              @endif
-                            </td>
-                            <td class='px-6 py-4 whitespace-nowrap'  style="position:relative;" data-field="created_at" >
-                              {{$user->created_at}}
-                            </td>
-                            <td class='px-6 py-4 whitespace-nowrap'  style="position:relative;" data-field="updated_at" >
-                              {{$user->updated_at}}
-                            </td> 
-                          </tr>
+                          <livewire:internal.user-row :key="$user->id" :user="$user"/>
                         @endforeach
                       </tbody>
                     </table>
